Split the app class code into multiple manage class files
